Added new markup. Refactored components export. Moved App component into components dir
