--- src/App.scss
@@ -1,8 +1,0 @@
-.container-center {
-    display: flex;
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    align-items: center;
-    justify-content: center;
-}
