+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export * from './App/App';
+export * from './Cell/Cell';
+export * from './CellGroup/CellGroup';
+export * from './Game/Game';
+export * from './NumBar/NumBar';
+export * from './Sudoku/Sudoku';
+export * from './Toolbar/Toolbar';
+export * from './Infobar/Infobar';
