+++ src/index.tsx
@@ -1,12 +1,12 @@
 import ReactDOM from 'react-dom';
 import './index.scss';
-import App from './App';
 import { Provider } from 'react-redux';
 import { applyMiddleware, createStore } from 'redux';
 import { reducers } from './reducers';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import React from 'react';
+import { App } from './components';
 
 const store = createStore(
     reducers,
